backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
alembic==1.13.1
openai==1.65.0
email-validator==2.1.0 
python-multipart==0.0.9 
psycopg2-binary==2.9.10
minio==7.1.14
aio-pika==8.0.0
kubernetes==30.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
kubernetes==30.1.0
loguru==0.7.2
redis==7.4.0
celery==5.6.3
tenacity==9.1.4